Add FleX, filter thickness, add sim controls, and change depth to flip buffers

+++ modules/gelly-fluid-sim/include/fluidsim/ISimScene.h
@@ -28,7 +28,7 @@ struct ObjectCreationParams {
 		 * the caller to ensure that this array is valid for the lifetime of the
 		 * create object function call.
 		 */
-		uint *indices;
+		ushort *indices;
 
 		uint vertexCount;
 		uint indexCount;
